Split `traits::error_reporting` to keep files smaller
